requirements.txt
tensorflow
opencv-python
numpy
scikit-learn
gradio
matplotlib
pillow
